feat: move inline svgs to icon components
